VIC-3523 updating FCC ALL menu item tests - now ALL test works for speaker and gears at the same time - added menu itm ALL (no CPU/battery) tests

+++ robot/fixture/helpware/files/testsquarewav.sh
@@ -1,4 +1,10 @@
 #!/bin/sh
+
+if [ "$1" == "" ]; then
+  /anki/menuman/testsquarewav.sh async &
+  exit 0
+fi
+
 killall -9 pcm && exit
 /etc/initscripts/anki-audio-init
 export LD_LIBRARY_PATH=/anki/menuman
